V15.17.01.23
Project Completion

+++ src/app/pages/products/products.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { CrudService } from 'src/app/services/crud.service';
 import { CartService } from 'src/app/services/cart.service';
-
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-products',
@@ -9,8 +9,13 @@ import { CartService } from 'src/app/services/cart.service';
   styleUrls: ['./products.component.scss']
 })
 export class ProductsComponent implements OnInit {
-
+  /**
+   *
+   */
   all_categories_list: any = [];
+  /**
+   * @description variable to store all the articles from
+   */
   all_articles_list:any = [];
 
   constructor(private crud_service: CrudService, public cart_service: CartService) { }
@@ -19,42 +24,90 @@ export class ProductsComponent implements OnInit {
     this.getAllCategories();
     this.getAllArticles();
   }
-
+  /**
+   * @description Function to fetch categories from database
+   * @returns list of categories array
+   */
   getAllCategories(){
     this.all_categories_list =  this.crud_service.getAllCategories();
+    return this.all_categories_list;
   }
-
+  /**
+   * @description Function to fetch articles from database
+   * @returns list of articles array
+   */
   getAllArticles(){
     this.all_articles_list =  this.crud_service.getAllArticles();
+    return this.all_articles_list;
   }
-
+  /**
+   * @description Function to Put item in the cart list
+   * @param article item object
+   * @returns boolean true if object is added, false if not
+   */
   addToCart(article: any){
-    this.cart_service.putInCart(article);
-    this.applyCss();
-    this.showSuccess();
+    let result = false;
+    if(article!=undefined || article!=null){
+      this.cart_service.putInCart(article);
+      this.showSuccess();
+      this.applyCss();
+      result = true;
+    }
+    return result;
   }
-
+ /**
+  * @description Function to increment item in the cart
+  * @param article item object
+  * @returns boolean true if object is added, false if not
+  */
   addItem(article:any){
-    this.cart_service.putInCart(article)
+    let result = false;
+    if(article!=undefined || article!=null){
+      this.cart_service.putInCart(article);
+      result = true;
+    }
+    return result;
   }
-
+ /**
+  * @description Function to decrement item from the cart
+  * @param article item object
+  * @returns boolean true if object is deleted, false if not
+  */
   minuItem(article:any){
-    this.cart_service.removeFromCart(article)
+    let result = false;
+    if(article!=null || article!=undefined){
+      this.cart_service.removeFromCart(article);
+      result = true;
+    }
+    return result;
   }
-
+  /**
+  * @description Function to add a CSS effect when something is aded to the cart
+  */
   applyCss(){
    let element: any = document.getElementById('lower_menu_cart');
-   element.classList.add('bounce');
-   setTimeout(()=>{
-    element.classList.remove("bounce");
-   }, 3000)
+    if(element!=null){
+      element.classList.add('bounce');
+      setTimeout(()=>{
+       element.classList.remove("bounce");
+      }, 3000);
+    }
   }
-
+  /**
+  * @description Function to notify the user that their item is added to the cart
+  */
   showSuccess() {
-    // this.toastr.success('Nice! You just added an item to the cart.');
+    Swal.fire({
+      icon: 'success',
+      title: 'Item added to cart.',
+      position: 'bottom-end',
+      showConfirmButton: false,
+      timer: 3000,
+      timerProgressBar: true,
+      didOpen: (toast) => {
+        toast.addEventListener('mouseenter', Swal.stopTimer)
+        toast.addEventListener('mouseleave', Swal.resumeTimer)
+      }
+    })
   }
-
-
-
-
 }
